requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.46.0
youtube-dl==2018.2.4
